Refactor models, add interfaces for factories & migration to db & tests for ParameterNormalizers

+++ DataAnalyzeAPI/Models/Entities/DataObject.cs
@@ -7,12 +7,12 @@ public class DataObject
 {
     [Key]
     [JsonIgnore]
-    public int Id { get; set; }
+    public long Id { get; set; }
 
     [Required]
     public string Name { get; set; } = string.Empty;
 
-    [JsonIgnore] public int DatasetId { get; set; }
+    [JsonIgnore] public long DatasetId { get; set; }
     [JsonIgnore] public Dataset Dataset { get; set; } = default!;
 
     public List<ParameterValue> Values { get; set; } = new();
